requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.1.3
jinja2==3.1.4
